+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 New features:
 * [OLMIS-3828](https://openlmis.atlassian.net/browse/OLMIS-3828): Added Service Account list UI.
+* [OLMIS-2695](https://openlmis.atlassian.net/browse/OLMIS-2695): Added Processing Schedule screens.
 
 Improvements: